Code Study homework 지금까지한거

+++ src/main/java/com/LostArk/lostark/homeworkItemRepository.java
@@ -2,5 +2,5 @@ package com.LostArk.lostark;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface homeworkItemRepository extends JpaRepository<homeworkItem, Integer> {
+public interface homeworkItemRepository extends JpaRepository<homeworkItem, Long> {
 }
